requirements.txt
argh==0.26.2
beautifulsoup4==4.8.1
bs4==0.0.1
Click==7.0
Flask==1.1.1
IMDbPY==6.8
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
numpy==1.17.3
pandas==0.25.3
pathtools==0.1.2
pprint==0.1
progress==1.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.11
watchdog==0.9.0
Werkzeug==0.16.0
